requirements.txt
oandapyV20>=0.7.2
python-dotenv>=1.0.0
pandas>=2.0.0
pandas-ta>=0.3.14b
requests>=2.31.0
flask>=3.0.0
python-dateutil>=2.8.2
matplotlib>=3.8.0
seaborn>=0.13.0
rich>=13.7.0
plotly>=5.18.0
